(doc nav) make expand/collapse tree button kb+sr friendly
- before, the arrow was a div with click handlers registered on it. now
it is an actual `button` tag, making keyboard nav work by default
- add some alt text for screen reader users
- reorder the DOM to make it more understandable that the
expand/collapse button is tied to the page just announced before it.

In a ideal world we would also have the page name in the toggle button
but it's not that easy to do right now. This will be good enough as a
first step.